Bugfix: RPC - simplified RPC and solved a bug.
The reply to the server was made 3 times in case of a correct
response by the device: the actual answer, the cancel answer and a second
cancel answare after the timeout.

We also simplified the whole procedure, in order to use only one
dictionary insted of a dictionary and a queue.

Finally, we employed a mutex when accessing the RPC dictionary.

Signed-off-by: nicofossa <fossanico@gmail.com>